Added PAGE.onPageAppear()
PAGE.onPageAppear() will be called when a page is about to be displayed.
This applies to the page now about to be shown when going back as well
as the menu page about to be shown.

This does not apply to the first time a page is being shown (initalize
has not yet been called reliably so anyone implementing an HTML page
should just use the initalize function) nor to the user returning to
the app.

Native pages should keep on using viewWillAppear: